Moved admin extensions to dedicated package
