+++ graphql-dxm-provider/src/main/java/org/jahia/modules/graphql/provider/dxm/admin/GqlAdminQuery.java
@@ -1,4 +1,4 @@
-package org.jahia.modules.graphql.provider.dxm.node;
+package org.jahia.modules.graphql.provider.dxm.admin;
 
 import graphql.annotations.annotationTypes.GraphQLDescription;
 import graphql.annotations.annotationTypes.GraphQLField;
